requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
autopep8==1.6.0
BareNecessities==0.2.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.6
cryptography==38.0.3
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.1.3
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-import-export==2.7.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
lxml==4.9.1
Mail==2.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyPDF3==1.0.6
python-bidi==0.4.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
self==2020.12.3
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
tzdata==2022.1
tzlocal==4.2
Unipath==1.1
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
